API review: move avocado.data_dir to avocado.core.data_dir
This is also a module that shouldn't be used directly by tests,
as tests themselves have attributes with the right data directories
already ready to use.

So, let's move it to avocado.core.
