+++ avocado/core/data_dir.py
@@ -39,7 +39,7 @@ from avocado.utils.data_structures import Borg
 from avocado.settings import settings
 
 
-_BASE_DIR = os.path.join(sys.modules[__name__].__file__, "..", "..")
+_BASE_DIR = os.path.join(sys.modules[__name__].__file__, "..", "..", "..")
 _BASE_DIR = os.path.abspath(_BASE_DIR)
 _IN_TREE_TESTS_DIR = os.path.join(_BASE_DIR, 'examples', 'tests')
