+++ selftests/all/unit/avocado/remote_unittest.py
@@ -5,7 +5,7 @@ import os
 
 from flexmock import flexmock, flexmock_teardown
 
-from avocado import data_dir
+from avocado.core import data_dir
 from avocado.core import remote
 from avocado.utils import archive
 from avocado.utils import remote as utils_remote
